+++ app/javascript/components/LoadProjectForm.js
@@ -2,33 +2,122 @@ import React from "react"
 import PropTypes from "prop-types"
 import { Container, Row, Col } from 'react-grid-system';
 
+import { Button, Input, Checkbox, Form } from 'semantic-ui-react'
+import '../../../dist/semantic.min.css';
+
+const FormTutor = props => (
+  <div className="formTutor">
+  <Form.Group inline>
+  <Form.Field>
+  <label>Dni</label>
+  <input placeholder='Dni' name={'dni_' + props.number} />
+  </Form.Field>
+  <Form.Field>
+  <label>Nombres</label>
+  <input placeholder='Nombres' name={'nombres_' + props.number} />
+  </Form.Field>
+  <Form.Field>
+  <label>Apellidos</label>
+  <input placeholder='Apellidos' name={'apellidos_' + props.number} />
+  </Form.Field>
+  </Form.Group>
+  <Form.Group inline>
+  <Form.Field>
+  <label>Institución</label>
+  <input placeholder='Institución' name={'institucion_' + props.number} />
+  </Form.Field>
+  <Form.Field>
+  <label>País</label>
+  <input placeholder='País' name={'pais_' + props.number} />
+  </Form.Field>
+  <Form.Field>
+  <label>Rol</label>
+  <select name={'rol_' + props.number}>
+  <option>Director</option>
+  <option>Tutor</option>
+  <option>Asesor</option>
+  </select>
+  </Form.Field>
+  </Form.Group>
+  </div>
+  )
+
+class LoadFile extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state =
+    {
+      tutors: [<FormTutor number={0} />]
+    }
+    this.tutorsNumber = 0;
+  }
+
+  onAddTutor = () => {
+    this.tutorsNumber += 1;
+    console.log(this.state);
+    this.setState(prevState => ({
+      tutors: [...prevState.tutors, <FormTutor number={this.tutorsNumber} />]
+    }))
+  }
+
+  render() {
+    return (
+      <div class="ui raised very padded text container segment">
+      <div class="content">
+      <h2 class="ui header">Cargar archivo</h2>
+      <div class="description">
+      <Form id="formProject" action='/file/load_post' method='POST' enctype="multipart/form-data">
+      <Form.Field>
+      <label>Titulo</label>
+      <input placeholder='Titulo' name='titulo' />
+      </Form.Field>
+      <Form.Field>
+      <label>Descripción</label>
+      <textarea placeholder='Descripción' name="descripcion" ></textarea>
+      </Form.Field>
+      <Form.Field>
+      <label>Información directores, tutores y asesores</label>
+
+      </Form.Field>
+
+      {this.state.tutors}
+
+      <Form.Field>
+      <Button onClick={this.onAddTutor} type='button'>Añadir</Button>
+      </Form.Field>
+      <Form.Field>
+      <label>Documentos</label>
+      </Form.Field>
+      <Form.Field>
+      <label>Propuesta</label>
+      <input type='file' name="file" />
+      </Form.Field>
+      <Form.Field>
+      <label>Documento de soporte</label>
+      <input type='file' name="supportFile" />
+      </Form.Field>
+      <Form.Field>
+      <label>Comentarios adicionales</label>
+      <textarea placeholder='Comentarios adicionales' name="descripcion" ></textarea>
+      </Form.Field>
+      <Form.Field>
+      <Button type='submit'>Cargar proyecto</Button>
+      </Form.Field>
+      </Form>
+      </div>
+      </div>
+      </div>
+      );
+  }
+}
+
 class LoadProjectForm extends React.Component {
   render () {
     return (
-    <Container>
-      <Row>
-        <Col sm={12}>
-          <Row>
-            <Col sm={4}>
-              <label>Nombre proyecto:</label>
-
-            </Col>
-          <Col sm={8}>
-            <input type="text" name="post[name]"></input>
-          </Col>
-
-          </Row>
-          
-        </Col>
-      </Row>
-      <Row>
-        <Col sm={12}>
-          <label>Descripcion proyecto:</label>
-          <textarea name="post[description]" rows="4" cols="50"></textarea>
-        </Col>
-      </Row>
-    </Container>
-    );
+      <div>
+      <LoadFile />
+      </div>
+      )
   }
 }
