+++ app/javascript/packs/application.js
@@ -7,8 +7,24 @@
 // To reference this file, add <%= javascript_pack_tag 'application' %> to the appropriate
 // layout file, like app/views/layouts/application.html.erb
 
-console.log('Hello World from Webpacker')
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from '../components/App';
+
+import '../../../dist/semantic.min.css';
+
 // Support component names relative to this directory:
 var componentRequireContext = require.context("components", true)
 var ReactRailsUJS = require("react_ujs")
 ReactRailsUJS.useContext(componentRequireContext)
+
+document.addEventListener('DOMContentLoaded', () => {
+	const wrapper = document.createElement('div');
+	wrapper.className = 'react-app';
+
+	ReactDOM.render(
+		<App />,
+		document.body.appendChild(wrapper)
+		);
+});
+
