+++ app/views/layouts/application.html.erb
@@ -9,7 +9,7 @@
     <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
     
     <%= javascript_pack_tag 'application' %>
-    <%= stylesheet_pack_tag 'login_form' %>
+    <%= stylesheet_pack_tag 'application' %>
   </head>
 
   <body>
