+++ app/views/sessions/new.html.erb
@@ -1,3 +1,3 @@
 <% provide(:title, "Login") %>
 
-<%= react_component("LoginForm") %>
+<!--<%= react_component("LoginForm") %>-->
